+++ src/Crowdin.Api/Glossaries/Glossary.cs
@@ -9,7 +9,7 @@ namespace Crowdin.Api.Glossaries
     public class Glossary
     {
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         [JsonProperty("name")]
         public string Name { get; set; }
@@ -18,10 +18,10 @@ namespace Crowdin.Api.Glossaries
         public int? GroupId { get; set; }
         
         [JsonProperty("userId")]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         
         [JsonProperty("terms")]
-        public int Terms { get; set; }
+        public long Terms { get; set; }
         
         [JsonProperty("languageId")]
         public string LanguageId { get; set; }
