+++ src/Crowdin.Api/Tasks/EnterprisePendingTaskCreateForm.cs
@@ -13,7 +13,7 @@ namespace Crowdin.Api.Tasks
     public class EnterprisePendingTaskCreateForm : AddTaskRequest
     {
         [JsonProperty("precedingTaskId")]
-        public int PrecedingTaskId { get; set; }
+        public long PrecedingTaskId { get; set; }
         
         [JsonProperty("type")]
         public TaskType Type { get; set; }
