+++ src/Crowdin.Api/TranslationMemory/TranslationMemory.cs
@@ -9,13 +9,13 @@ namespace Crowdin.Api.TranslationMemory
     public class TranslationMemory
     {
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         [JsonProperty("groupId")]
         public int? GroupId { get; set; }
         
         [JsonProperty("userId")]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         
         [JsonProperty("name")]
         public string Name { get; set; }
